chore(taro-rn): rebase 代码,调整结构
